[MIG] pos_product_template: Migration to 15.0
