+++ pos_product_template/README.rst
@@ -7,7 +7,7 @@ POS - Product Template
    !! This file is generated by oca-gen-addon-readme !!
    !! changes will be overwritten.                   !!
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! source digest: sha256:4a1b25e7f68d82e67c53ff742cadd7fdbd84e7af503e639c4e1972a713a18772
+   !! source digest: sha256:db57e93b7dc8c62c24c51e03b823c331eac215f8938027e6c0949f97e7f6e202
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 .. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
@@ -20,7 +20,7 @@ POS - Product Template
     :target: https://github.com/OCA/pos/tree/15.0/pos_product_template
     :alt: OCA/pos
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/pos-15.0/pos-15.0-pos_product_template
+    :target: https://translation.odoo-community.org/projects/pos-15-0/pos-15-0-pos_product_template
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runboat-Try%20me-875A7B.png
     :target: https://runboat.odoo-community.org/builds?repo=OCA/pos&target_branch=15.0
@@ -86,6 +86,10 @@ Contributors
 * `Ooops <https://www.ooops404.com>`_:
 
    * Giovanni Serra <giovanni@gslab.it>
+
+* `Aures Tic <https://aurestic.es>`_:
+
+   * Jose Zambudio <jose@aurestic.es>
 
 Funders
 -------
