+++ pos_product_template/static/src/css/ppt.css
@@ -1,65 +1,14 @@
-/******************************************************************************
-    Point Of Sale - Product Template module for Odoo
-    Copyright (C) 2014-Today Akretion (http://www.akretion.com)
-    @author Sylvain LE GAL (https://twitter.com/legalsylvain)
-
-    Some part of code from Odoo Project,
-    Copyright (C) 2004-Today Odoo SA. (<http://odoo.com>).
-
-    This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU Affero General Public License as
-    published by the Free Software Foundation, either version 3 of the
-    License, or (at your option) any later version.
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU Affero General Public License for more details.
-
-    You should have received a copy of the GNU Affero General Public License
-    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-******************************************************************************/
-
 /*  ********* Variant Selector PopUp  ************************************** */
-.pos .modal-dialog .popup-select-variant {
-    margin-left: 0;
-    margin-top: 0;
-    left: 5%;
-    width: 90%;
-    top: 5%;
-    height: 85%;
+.pos .modal-dialog .select-variant-products {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    flex-wrap: wrap;
+    gap: 2px;
 }
 
-.pos .modal-dialog .popup-select-variant .variant-title {
-    height: 8%;
-}
-
-.pos .modal-dialog .popup-select-variant .variant-title #variant-popup-cancel {
-    margin: -10px 3px 3px 3px;
-}
-
-.pos .modal-dialog .popup-select-variant .container-attribute-list {
-    width: 100%;
-    height: 40%;
-    overflow: hidden;
-    overflow-y: auto;
-    border-bottom: 1px dashed #444;
-    -webkit-overflow-scrolling: touch;
-}
-
-.pos .modal-dialog .popup-select-variant .container-variant-list {
-    width: 100%;
-    height: 50%;
-    overflow: hidden;
-    overflow-y: auto;
-
-    -webkit-overflow-scrolling: touch;
-}
-
-.pos .variant-list {
-    padding: 10px;
-    text-align: left;
-    -webkit-transform: translate3d(0, 0, 0);
+.pos .modal-dialog .select-variant-products article {
+    font-size: 12px;
 }
 
 /*  ********* Attribut Display  ******************************************** */
@@ -133,103 +82,3 @@
     font-size: 9px;
     margin-right: 3px;
 }
-
-/*  ********* Variant Display  ********************************************* */
-/* This part is a copy paste from Odoo Project, with some values changed     */
-.pos .variant {
-    position: relative;
-    vertical-align: top;
-    display: inline-block;
-    line-height: 100px;
-    font-size: 11px;
-    margin: 5px !important;
-    width: 240px;
-    height: 60px;
-    background: #fff;
-    border: 1px solid #e2e2e2;
-    border-radius: 3px;
-    border-bottom-width: 3px;
-    overflow: hidden;
-    cursor: pointer;
-}
-
-.pos .variant .variant-header {
-    position: relative;
-    width: 240px;
-    height: 30px;
-    background: white;
-    text-align: center;
-}
-
-.pos .variant .price-tag {
-    position: absolute;
-    top: 2px;
-    right: 2px;
-    vertical-align: top;
-    color: white;
-    line-height: 13px;
-    background: #7f82ac;
-    padding: 2px 5px;
-    border-radius: 2px;
-}
-
-.pos .variant .variant-name {
-    font-weight: normal;
-    position: absolute;
-    -webkit-box-sizing: border-box;
-    -moz-box-sizing: border-box;
-    -ms-box-sizing: border-box;
-    box-sizing: border-box;
-    bottom: 0;
-    top: auto;
-    line-height: 14px;
-    width: 100%;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    background: -webkit-linear-gradient(
-        -90deg,
-        rgba(255, 255, 255, 0),
-        rgba(255, 255, 255, 1),
-        rgba(255, 255, 255, 1)
-    );
-    background: -moz-linear-gradient(
-        -90deg,
-        rgba(255, 255, 255, 0),
-        rgba(255, 255, 255, 1),
-        rgba(255, 255, 255, 1)
-    );
-    background: -ms-linear-gradient(
-        -90deg,
-        rgba(255, 255, 255, 0),
-        rgba(255, 255, 255, 1),
-        rgba(255, 255, 255, 1)
-    );
-    /* troublesome in latest webkit
-    background:         linear-gradient(-90deg,rgba(255,255,255,0),rgba(255,255,255,1), rgba(255,255,255,1));
-    */
-    /*background:#FFF;*/
-    padding: 3px;
-    padding-top: 15px;
-}
-
-.pos .button.disabled {
-    background: #e2e2e2 !important;
-    border: solid 1px #bebebe !important;
-    color: grey !important;
-    opacity: 0.5;
-    cursor: not-allowed;
-}
-
-/*.pos .variant-list-scroller{*/
-/*    -webkit-box-sizing: border-box;*/
-/*       -moz-box-sizing: border-box;*/
-/*        -ms-box-sizing: border-box;*/
-/*            box-sizing: border-box;*/
-/*    width:50%;*/
-/*    height:50%;*/
-/*    overflow: hidden;*/
-/*    overflow-y: auto;*/
-/*    -webkit-overflow-scrolling: touch;*/
-/*    -webkit-transform: translate3d(0,0,0);*/
-
-/*}*/
